refactor: make utils directory
