+++ src/libs/index.ts
@@ -1,2 +1,1 @@
 export * from "./notion/index.js";
-export * from "./utils.js";
