+++ src/libs/notion/databases.ts
@@ -1,6 +1,6 @@
 import { isFullDatabase } from "@notionhq/client";
 
-import { callAPIWithBackOff } from "../utils.js";
+import { callAPIWithBackOff } from "../utils/api.js";
 
 import { notion } from "./auth.js";
 import { convertResponseToRichText } from "./richText/richText.js";
